Move '_helpers.get_scoped_connection' -> 'connection.
